#2 - 🎨 Make post components.
